+++ keep/rulesengine/rulesengine.py
@@ -1,7 +1,6 @@
 import itertools
 import json
 import logging
-import re
 
 import celpy
 import chevron
@@ -13,6 +12,7 @@ from keep.api.core.db import get_rules as get_rules_db
 from keep.api.core.elastic import ElasticClient
 from keep.api.models.alert import AlertDto, AlertSeverity, AlertStatus
 from keep.api.models.group import GroupDto
+from keep.api.utils.cel_utils import preprocess_cel_expression
 
 
 class RulesEngine:
@@ -320,46 +320,6 @@ class RulesEngine:
         return group_payload
 
     @staticmethod
-    def preprocess_cel_expression(cel_expression: str) -> str:
-        """Preprocess CEL expressions to replace string-based comparisons with numeric values where applicable."""
-
-        # Construct a regex pattern that matches any severity level or other comparisons
-        # and accounts for both single and double quotes as well as optional spaces around the operator
-        severities = "|".join(
-            [f"\"{severity.value}\"|'{severity.value}'" for severity in AlertSeverity]
-        )
-        pattern = rf"(\w+)\s*([=><!]=?)\s*({severities})"
-
-        def replace_matched(match):
-            field_name, operator, matched_value = (
-                match.group(1),
-                match.group(2),
-                match.group(3).strip("\"'"),
-            )
-
-            # Handle severity-specific replacement
-            if field_name.lower() == "severity":
-                severity_order = next(
-                    (
-                        severity.order
-                        for severity in AlertSeverity
-                        if severity.value == matched_value.lower()
-                    ),
-                    None,
-                )
-                if severity_order is not None:
-                    return f"{field_name} {operator} {severity_order}"
-
-            # Return the original match if it's not a severity comparison or if no replacement is necessary
-            return match.group(0)
-
-        modified_expression = re.sub(
-            pattern, replace_matched, cel_expression, flags=re.IGNORECASE
-        )
-
-        return modified_expression
-
-    @staticmethod
     def filter_alerts(alerts: list[AlertDto], cel: str):
         """This function filters alerts according to a CEL
 
@@ -384,7 +344,7 @@ class RulesEngine:
             logger.debug("No CEL expression provided")
             return alerts
         # preprocess the cel expression
-        cel = RulesEngine.preprocess_cel_expression(cel)
+        cel = preprocess_cel_expression(cel)
         ast = env.compile(cel)
         prgm = env.program(ast)
         filtered_alerts = []
